requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pytz>=2023.3
discord.py>=2.3.0
gspread>=6.0.0
